moved createContext to a different file
